Attempted to Improved code functionality
I went through and tried making the code more clear, and safe by making the list less mutable. However, files can no longer be written to at the moment. Will fix when I get a chance. Just commiting to have it saved.

+++ src/file_management.rs
@@ -68,7 +68,9 @@ pub fn create_file(path_to_file : &PathBuf, file_name : String) {
 
     let file_path = path_to_file.join(file);
 
-    fs::write(&file_path, "[]").expect("Could not write to file");
+    if !path_to_file.exists() {
+        fs::write(&file_path, "[]").expect("Could not write to file");
+    }
 
 }
